+++ dyn_debug/fun_call.cc
@@ -5,7 +5,7 @@
 void show_elf_fun_call(pid_t pid, char* elf_fun_name)
 {
     char fun_code[0x1000];
-    unsigned long long fun_start_addr, fun_end_addr, fun_size;
+    u64 fun_start_addr, fun_end_addr, fun_size;
     string fun_name;
 
     fun_start_addr = get_elf_fun_addr(elf_fun_name);
